# lib/api-zod/src/generated/types/configLojasInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * INFINITYPHONE ESTOQUE - Phone inventory management API
 * OpenAPI spec version: 0.1.0
 */
import type { ConfigLojasInputLojasItem } from './configLojasInputLojasItem';

export interface ConfigLojasInput {
  lojas: ConfigLojasInputLojasItem[];
}