# lib/api-zod/src/generated/types/movimentacaoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * INFINITYPHONE ESTOQUE - Phone inventory management API
 * OpenAPI spec version: 0.1.0
 */
import type { MovimentacaoInputLojaNumero } from './movimentacaoInputLojaNumero';
import type { MovimentacaoInputTipo } from './movimentacaoInputTipo';

export interface MovimentacaoInput {
  produtoId: number;
  lojaNumero: MovimentacaoInputLojaNumero;
  tipo: MovimentacaoInputTipo;
  /** @minimum 1 */
  quantidade: number;
  observacao?: string;
}