# lib/api-zod/src/generated/types/produtoComEstoque.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * INFINITYPHONE ESTOQUE - Phone inventory management API
 * OpenAPI spec version: 0.1.0
 */
import type { EstoquePorLoja } from './estoquePorLoja';
import type { ProdutoComEstoqueStatusNovoUsado } from './produtoComEstoqueStatusNovoUsado';

export interface ProdutoComEstoque {
  id: number;
  modelo: string;
  marca: string;
  cor: string;
  armazenamentoGb: number;
  ramGb: number;
  precoCusto: number;
  precoVenda: number;
  statusNovoUsado: ProdutoComEstoqueStatusNovoUsado;
  dataCadastro: Date;
  estoqueTotal: number;
  estoquePorLoja: EstoquePorLoja[];
  chegouHoje: boolean;
  /** @nullable */
  imagemUrl?: string | null;
  /** @nullable */
  imagemFile?: string | null;
}