# lib/api-zod/src/generated/types/valorLoja.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * INFINITYPHONE ESTOQUE - Phone inventory management API
 * OpenAPI spec version: 0.1.0
 */

export interface ValorLoja {
  lojaNumero: number;
  valorTotal: number;
  quantidadeTotal: number;
}